+++ src/perception/ObjectDatabase.cpp
@@ -1,4 +1,4 @@
-#include <aikido/perception/ObjectDatabase.hpp>
+#include "aikido/perception/ObjectDatabase.hpp"
 
 #include <dart/common/Console.hpp>
 #include <dart/common/LocalResourceRetriever.hpp>
@@ -33,11 +33,12 @@ ObjectDatabase::ObjectDatabase(
 void ObjectDatabase::getObjectByKey(
     const std::string& objectKey,
     std::string& objectName,
-    dart::common::Uri& objectResource) const
+    dart::common::Uri& objectResource,
+    Eigen::Isometry3d& objectOffset) const
 {
   // Get name of object and pose for a given tag ID
-  YAML::Node obj_node = mObjData[objectKey];
-  if (!obj_node)
+  YAML::Node objectNode = mObjData[objectKey];
+  if (!objectNode)
   {
     throw std::runtime_error("[ObjectDatabase] Error: invalid object key");
   }
@@ -45,7 +46,7 @@ void ObjectDatabase::getObjectByKey(
   // Convert resource field
   try
   {
-    objectResource.fromString(obj_node["resource"].as<std::string>());
+    objectResource.fromString(objectNode["resource"].as<std::string>());
   }
   catch (const YAML::ParserException& ex)
   {
@@ -56,12 +57,23 @@ void ObjectDatabase::getObjectByKey(
   // Convert name field
   try
   {
-    objectName = obj_node["name"].as<std::string>();
+    objectName = objectNode["name"].as<std::string>();
   }
   catch (const YAML::ParserException& ex)
   {
     throw std::runtime_error(
         "[ObjectDatabase] Error in converting [name] field");
+  }
+
+  // Convert offset field
+  try
+  {
+    objectOffset = objectNode["offset"].as<Eigen::Isometry3d>();
+  }
+  catch (const YAML::ParserException& ex)
+  {
+    throw std::runtime_error(
+        "[ObjectDatabase] Error in converting [offset] field");
   }
 }