+++ src/perception/PoseEstimatorModule.cpp
@@ -1,4 +1,4 @@
-#include <aikido/perception/RcnnPoseModule.hpp>
+#include "aikido/perception/PoseEstimatorModule.hpp"
 
 #include <Eigen/Geometry>
 #include <dart/utils/urdf/DartLoader.hpp>
@@ -6,13 +6,14 @@
 #include <ros/topic.h>
 #include <visualization_msgs/Marker.h>
 #include <visualization_msgs/MarkerArray.h>
-#include <aikido/io/CatkinResourceRetriever.hpp>
-#include <aikido/perception/shape_conversions.hpp>
+#include "aikido/io/CatkinResourceRetriever.hpp"
+#include "aikido/perception/shape_conversions.hpp"
 
 namespace aikido {
 namespace perception {
 
-RcnnPoseModule::RcnnPoseModule(
+//=============================================================================
+PoseEstimatorModule::PoseEstimatorModule(
     ros::NodeHandle nodeHandle,
     std::string markerTopic,
     std::shared_ptr<ObjectDatabase> configData,
@@ -30,7 +31,8 @@ RcnnPoseModule::RcnnPoseModule(
   // Do nothing
 }
 
-bool RcnnPoseModule::detectObjects(
+//=============================================================================
+bool PoseEstimatorModule::detectObjects(
     const aikido::planner::WorldPtr& env,
     ros::Duration timeout,
     ros::Time timestamp)
@@ -42,17 +44,17 @@ bool RcnnPoseModule::detectObjects(
       = ros::topic::waitForMessage<visualization_msgs::MarkerArray>(
           mMarkerTopic, mNodeHandle, timeout);
 
-  // Making sure the Message from RcnnPose is non empty
+  // Making sure the Message from a pose estimator is non empty
   if (marker_message == nullptr)
   {
-    dtwarn << "[RcnnPoseModule::detectObjects] nullptr Marker Message "
+    dtwarn << "[PoseEstimatorModule::detectObjects] nullptr Marker Message "
            << mMarkerTopic << std::endl;
     return false;
   }
 
-  if (marker_message->markers.size() == 0)
+  if (marker_message->markers.empty())
   {
-    dtwarn << "[RcnnPoseModule::detectObjects] No markers on topic "
+    dtwarn << "[PoseEstimatorModule::detectObjects] No markers on topic "
            << mMarkerTopic << std::endl;
     return false;
   }
@@ -85,10 +87,11 @@ bool RcnnPoseModule::detectObjects(
 
     std::string obj_name;
     dart::common::Uri obj_resource;
+    Eigen::Isometry3d obj_offset;
     ros::Time t0 = ros::Time(0);
 
-    // get the object name and resource from database
-    mConfigData->getObjectByKey(obj_key, obj_name, obj_resource);
+    // get the object name, resource, and offset from database by objectKey
+    mConfigData->getObjectByKey(obj_key, obj_name, obj_resource, obj_offset);
 
     tf::StampedTransform transform;
     try
@@ -101,7 +104,7 @@ bool RcnnPoseModule::detectObjects(
     }
     catch (const tf::ExtrapolationException& ex)
     {
-      dtwarn << "[RcnnPoseModule::detectObjects] TF timestamp is "
+      dtwarn << "[PoseEstimatorModule::detectObjects] TF timestamp is "
                 "out-of-date compared to marker timestamp "
              << ex.what() << std::endl;
       continue;
@@ -114,7 +117,8 @@ bool RcnnPoseModule::detectObjects(
     Eigen::Isometry3d frame_pose
         = aikido::perception::convertStampedTransformToEigen(transform);
     // Compose to get actual skeleton pose
-    Eigen::Isometry3d obj_pose = frame_pose * marker_pose;
+    Eigen::Isometry3d obj_pose = frame_pose * marker_pose * obj_offset;
+
     Eigen::Isometry3d link_offset = mReferenceLink->getWorldTransform();
     obj_pose = link_offset * obj_pose;
 
@@ -124,7 +128,7 @@ bool RcnnPoseModule::detectObjects(
     // Check if skel in World
     // If there is, update its pose
     // If not, add skeleton to env
-    // RcnnPose module should provide the unique obj_id per object
+    // A pose estimator module should provide the unique obj_id per object
     if (env_skeleton == nullptr)
     {
       is_new_obj = true;
@@ -133,8 +137,9 @@ bool RcnnPoseModule::detectObjects(
 
       if (!obj_skeleton)
       {
-        dtwarn << "[RcnnPoseModule::detectObjects] Failed to load skeleton "
-               << "for URI " << obj_resource.toString() << std::endl;
+        dtwarn
+            << "[PoseEstimatorModule::detectObjects] Failed to load skeleton "
+            << "for URI " << obj_resource.toString() << std::endl;
         continue;
       }
       obj_skeleton->setName(obj_id);
@@ -152,7 +157,7 @@ bool RcnnPoseModule::detectObjects(
     }
     else
     {
-      dtwarn << "[RcnnPoseModule::detectObjects] Skeleton " << obj_name
+      dtwarn << "[PoseEstimatorModule::detectObjects] Skeleton " << obj_name
              << " has 0 DOFs! \n";
       continue;
     }
@@ -163,7 +168,7 @@ bool RcnnPoseModule::detectObjects(
 
     if (freejtptr == nullptr)
     {
-      dtwarn << "[RcnnPoseModule::detectObjects] Could not cast the joint "
+      dtwarn << "[PoseEstimatorModule::detectObjects] Could not cast the joint "
                 "of the body to a Free Joint so ignoring the object "
              << obj_name << std::endl;
       continue;
@@ -181,7 +186,7 @@ bool RcnnPoseModule::detectObjects(
 
   if (!any_detected)
   {
-    dtwarn << "[RcnnPoseModule::detectObjects] No marker up-to-date with "
+    dtwarn << "[PoseEstimatorModule::detectObjects] No marker up-to-date with "
               "timestamp parameter"
            << std::endl;
     return false;